[#3294] fix PR comments and clean up code add casts to AggregationFunctionTest.cpp

+++ nes-runtime/tests/UnitTests/Execution/Aggregation/AggregationFunctionTest.cpp
@@ -48,7 +48,7 @@ TEST_F(AggregationFunctionTest, scanEmitPipelineSum) {
     auto sumValue = Aggregation::SumAggregationValue<int64_t>();
     auto memref = Nautilus::Value<Nautilus::MemRef>((int8_t*) &sumValue);
 
-    auto incomingValue = Nautilus::Value<Nautilus::Int64>( 1L);
+    auto incomingValue = Nautilus::Value<Nautilus::Int64>((int64_t) 1);
     // test lift
     sumAgg.lift(memref, incomingValue);
     ASSERT_EQ(sumValue.sum, 1);
@@ -79,7 +79,7 @@ TEST_F(AggregationFunctionTest, scanEmitPipelineCount) {
     auto countValue = Aggregation::CountAggregationValue<uint64_t>();
     auto memref = Nautilus::Value<Nautilus::MemRef>((int8_t*) &countValue);
 
-    auto incomingValue = Nautilus::Value<Nautilus::Int64>(1L);
+    auto incomingValue = Nautilus::Value<Nautilus::Int64>((int64_t) 1);
     // test lift
     countAgg.lift(memref, incomingValue);
     ASSERT_EQ(countValue.count, (uint64_t) 1);
@@ -140,8 +140,8 @@ TEST_F(AggregationFunctionTest, scanEmitPipelineMin) {
     auto minValue = Aggregation::MinAggregationValue<int64_t>();
     auto memref = Nautilus::Value<Nautilus::MemRef>((int8_t*) &minValue);
     auto incomingValueFive = Nautilus::Value<Nautilus::Int64>((int64_t) 5);
-    auto incomingValueTen = Nautilus::Value<Nautilus::Int64>(10L);
-    auto incomingValueOne = Nautilus::Value<Nautilus::Int64>(1L);
+    auto incomingValueTen = Nautilus::Value<Nautilus::Int64>((int64_t) 10);
+    auto incomingValueOne = Nautilus::Value<Nautilus::Int64>((int64_t) 1);
     auto incomingValueTwo = Nautilus::Value<Nautilus::Int64>((int64_t) 2);
 
     // lift value in minAgg with an initial value of 5, thus the current min should be 5
@@ -197,9 +197,9 @@ TEST_F(AggregationFunctionTest, scanEmitPipelineMax) {
     auto maxValue = Aggregation::MaxAggregationValue<int64_t>();
     auto memref = Nautilus::Value<Nautilus::MemRef>((int8_t*) &maxValue);
     auto incomingValueFive = Nautilus::Value<Nautilus::Int64>((int64_t) 5);
-    auto incomingValueTen = Nautilus::Value<Nautilus::Int64>(10L);
-    auto incomingValueOne = Nautilus::Value<Nautilus::Int64>(1L);
-    auto incomingValueFifteen = Nautilus::Value<Nautilus::Int64>(15L);
+    auto incomingValueTen = Nautilus::Value<Nautilus::Int64>((int64_t) 10);
+    auto incomingValueOne = Nautilus::Value<Nautilus::Int64>((int64_t) 1);
+    auto incomingValueFifteen = Nautilus::Value<Nautilus::Int64>((int64_t) 15);
 
     // lift value in maxAgg with an initial value of 5, thus the current min should be 5
     maxAgg.lift(memref, incomingValueFive);
